Move to advisory hooks
