+++ frontend/src/hooks/useAccess.js
@@ -11,13 +11,6 @@ export default function useAccess() {
     [checkAccess],
   );
 
-  // Returns the createdByRole/modifiedByRole value for the user's highest advisory role.
-  const getUserAdvisoryRole = useCallback(() => {
-    if (hasAnyRole([ROLES.ADVISORY_APPROVER])) return "approver";
-    if (hasAnyRole([ROLES.ADVISORY_SUBMITTER])) return "submitter";
-    return "contributor";
-  }, [hasAnyRole]);
-
   return {
     roles,
     checkAccess,
@@ -25,6 +18,5 @@ export default function useAccess() {
     logOut,
     isAuthenticated,
     hasAnyRole,
-    getUserAdvisoryRole,
   };
 }
